Changes for 1.21.3
- Updated a few mixin component for 1.21.3
- Implemented antlr-runtime due to neoforge for somem reason missing TokenSource class from antlr
- Changes POW AssignOperation from a^b to b^a

+++ src/main/java/yuria/sul/ast/item/ItemProperty.java
@@ -6,9 +6,9 @@ import yuria.sul.ast.AssignOperation;
 public class ItemProperty {
     public final Item item;
     public final AssignOperation assignOperation;
-    public final Integer doOpBy;
+    public final Long doOpBy;
 
-    public ItemProperty(Item item, AssignOperation operation, Integer doOpBy) {
+    public ItemProperty(Item item, AssignOperation operation, Long doOpBy) {
         this.assignOperation = operation;
         this.doOpBy = doOpBy;
         this.item = item;
