+++ src/main/java/yuria/sul/ast/item/UnfiliteredItemProperty.java
@@ -5,9 +5,9 @@ import yuria.sul.ast.ComparisonOperator;
 
 public class UnfiliteredItemProperty {
     public final AssignOperation assignOperation;
-    public final int assingBy;
+    public final long assingBy;
 
-    public UnfiliteredItemProperty(AssignOperation assignOperation, int assingBy)
+    public UnfiliteredItemProperty(AssignOperation assignOperation, long assingBy)
     {
         this.assignOperation = assignOperation;
         this.assingBy = assingBy;
@@ -17,7 +17,7 @@ public class UnfiliteredItemProperty {
         public final ComparisonOperator comparisonOperator;
         public final int compareBy;
 
-        public ItemPropertySize(AssignOperation assignOperation, int assingBy, ComparisonOperator comparisonOperator, int compareBy) {
+        public ItemPropertySize(AssignOperation assignOperation, long assingBy, ComparisonOperator comparisonOperator, int compareBy) {
             super(assignOperation, assingBy);
             this.comparisonOperator = comparisonOperator;
             this.compareBy = compareBy;
@@ -26,7 +26,7 @@ public class UnfiliteredItemProperty {
 
     public static class ItemPropertyId extends UnfiliteredItemProperty {
         public final String toMatch;
-        public ItemPropertyId(AssignOperation assignOperation, int assingBy, String toMatch) {
+        public ItemPropertyId(AssignOperation assignOperation, long assingBy, String toMatch) {
             super(assignOperation, assingBy);
             this.toMatch = toMatch;
         }
@@ -36,12 +36,12 @@ public class UnfiliteredItemProperty {
         public final String tag;
         public boolean isRegex = false;
 
-        public TagItemProperty(AssignOperation assignOperation, int assingBy, String tag, boolean isRegex) {
+        public TagItemProperty(AssignOperation assignOperation, long assingBy, String tag, boolean isRegex) {
             super(assignOperation, assingBy);
             this.tag = tag;
             this.isRegex = isRegex;
         }
-        public TagItemProperty(AssignOperation assignOperation, int assingBy, String tag) {
+        public TagItemProperty(AssignOperation assignOperation, long assingBy, String tag) {
             super(assignOperation, assingBy);
             this.tag = tag;
         }
